% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_query.R
\name{log_query}
\alias{log_query}
\title{get query}
\usage{
log_query(conn, query, namespace = "global")
}
\arguments{
\item{conn}{connection}

\item{query}{query}

\item{namespace}{logger namespace to use}
}
\value{
data.frame
}
\description{
get query
}
